requirements.txt
flask
flask-sqlalchemy
pymysql
flask-marshmallow
marshmallow-sqlalchemy